backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.35.3
huggingface_hub>=0.25.0
